+++ lista/13/lucas.js
@@ -3,7 +3,7 @@
 const prompt = require('prompt-sync')({})
 
 const lerSalario = () => {
-   let salario = parseInt(prompt("Digite o salário: ")) 
+   let salario = parseFloat(prompt("Digite o salário: ")) 
    console.log ("Seu novo salário, com 15% de aumento é de: ", salario*0.15+salario)
 
 
@@ -12,7 +12,7 @@ const lerSalario = () => {
 lerSalario();
 
 const novoSalario = () => {
-    let salario = parseInt(prompt("Digite o salário: "))
+    let salario = parseFloat(prompt("Digite o salário: "))
     console.log("VOCÊ TEVE UM AUMENTO, SEU NOVO SALÁRIO É DE: R$", salario*1.15)
 }
 novoSalario();
